1. Changed the input and output device indexes from long to int for the Windows settings dialog. 2. Changed Installed member function to IsInstalled for all of the players and changed then corresponding data member from installed to mInstalled. 3. Added code for to use the portmidi device selection dialog in the project. 4. Made cosmetic changes in the portmidi player class. 5. Changed the configuration code to use strings instead of char* in several places. 6. Removed the configuration code that read and wrote long values. 7. Changed JZConfiguration::Check to tokenize the passed line using a new function in Globals.cpp. 8. Changed some I/O in the configuration class from C-style to C++ stream. More to come in this area.
git-svn-id: https://svn.code.sf.net/p/jazzplusplus/code/trunk/jazz@579 d001d500-7bcc-4352-ac0f-06d6f3fd1383

+++ src/MidiDeviceDialog.cpp
@@ -40,7 +40,7 @@ END_EVENT_TABLE()
 //-----------------------------------------------------------------------------
 JZMidiDeviceDialog::JZMidiDeviceDialog(
   const vector<pair<string, int> >& MidiDevices,
-  long& DeviceIndex,
+  int& DeviceIndex,
   wxWindow* pParent,
   const wxString& Title)
   : wxDialog(pParent, wxID_ANY, Title),
